Datamodel/user authentication

+++ src/main/java/com/virtu_stock/Security/Authentication/AuthController.java
@@ -15,6 +15,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
@@ -26,6 +27,7 @@ import com.virtu_stock.Mail.OTP.OTPService;
 import com.virtu_stock.Security.Authentication.DTO.AuthRequest;
 import com.virtu_stock.Security.Authentication.DTO.ResetPasswordRequestDTO;
 import com.virtu_stock.Security.JWT.JWTUtil;
+import com.virtu_stock.User.CustomUserDetails;
 import com.virtu_stock.User.CustomUserDetailsService;
 import com.virtu_stock.User.User;
 import com.virtu_stock.User.UserResponseDTO;
@@ -51,48 +53,62 @@ public class AuthController {
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody AuthRequest request) {
         try {
-            // Authenticate user credentials
             authenticationManager.authenticate(
                     new UsernamePasswordAuthenticationToken(request.getEmail(), request.getPassword()));
 
             final var userDetails = userDetailsService.loadUserByUsername(request.getEmail());
-            final String jwt = jwtUtil.generateToken(userDetails);
-            return ResponseEntity.ok(Map.of("token", jwt));
+            CustomUserDetails customUserDetails = (CustomUserDetails) userDetails;
+            final String jwt = jwtUtil.generateToken(customUserDetails);
+            return ResponseEntity.ok(Map.of("virtustock-token", jwt));
 
         } catch (BadCredentialsException ex) {
             return ResponseEntity
                     .status(HttpStatus.UNAUTHORIZED)
-                    .body(Map.of("error", "Invalid email or password"));
+                    .body(Map.of("message", "Invalid email or password"));
 
         } catch (DisabledException ex) {
             return ResponseEntity
                     .status(HttpStatus.FORBIDDEN)
-                    .body(Map.of("error", "User account is disabled"));
+                    .body(Map.of("message", "User account is disabled"));
 
         } catch (UsernameNotFoundException ex) {
             return ResponseEntity
                     .status(HttpStatus.NOT_FOUND)
-                    .body(Map.of("error", "User not found"));
+                    .body(Map.of("message", "User not found"));
 
         } catch (Exception ex) {
             return ResponseEntity
                     .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "An unexpected error occurred"));
+                    .body(Map.of("message", "An unexpected error occurred. Please try again or contact Support Team"));
         }
     }
 
     @PostMapping("/register")
-    public ResponseEntity<?> registerUser(@Valid @RequestBody User user, BindingResult result) {
+    public ResponseEntity<?> registerUser(@RequestHeader(value = "x-otp-verify-token", required = false) String token,
+            @Valid @RequestBody User user, BindingResult result) {
+
         try {
             if (result.hasErrors()) {
                 Map<String, String> errors = new HashMap<>();
                 result.getFieldErrors().forEach(error -> errors.put(error.getField(), error.getDefaultMessage()));
-                return ResponseEntity.badRequest().body(Map.of("errors", errors));
+                return ResponseEntity.badRequest().body(Map.of("message", errors));
+            }
+            if (token == null || token.isBlank()) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body(Map.of("message", "Invalid Request. Please verify your email first"));
+            }
+
+            String emailFromToken = jwtUtil.validateOTPToken(token, OTPPurpose.SIGN_UP);
+
+            if (emailFromToken == null || !emailFromToken.equals(user.getEmail())) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body(Map.of("message", "Invalid email."));
             }
 
             if (!otpService.isEmailVerified(user.getEmail(), OTPPurpose.SIGN_UP)) {
                 return ResponseEntity.badRequest().body(
-                        Map.of("error", "Please verify your email before registration."));
+                        Map.of("message", "Please verify your email before registration."));
             }
 
             User savedUser = userService.registerUser(user);
@@ -104,15 +120,15 @@ public class AuthController {
 
         } catch (ResponseStatusException ex) {
             return ResponseEntity.status(ex.getStatusCode())
-                    .body(Map.of("error", ex.getReason()));
+                    .body(Map.of("message", ex.getReason()));
 
         } catch (IllegalArgumentException ex) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", ex.getMessage()));
+                    .body(Map.of("message", ex.getMessage()));
 
         } catch (Exception ex) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "An unexpected error occurred"));
+                    .body(Map.of("message", "An unexpected error occurred"));
         }
     }
 
@@ -121,41 +137,43 @@ public class AuthController {
         String email = request.get("email");
         if (email == null || email.isBlank()) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "Email is required"));
+                    .body(Map.of("message", "Email is required"));
         }
 
         String emailRegex = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$";
         if (!email.matches(emailRegex)) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "Invalid email format"));
+                    .body(Map.of("message", "Invalid email format"));
         }
 
         if (!mailService.hasMXRecord(email)) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "Invalid email domain. Please enter a valid email address."));
+                    .body(Map.of("message", "Invalid email domain. Please enter a valid email address."));
         }
 
         if (userService.existsByEmail(email)) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "User Already exists"));
+                    .body(Map.of("message", "User Already exists"));
         }
 
         // if (!emailVerificationService.verifyEmail(email)) {
         // return ResponseEntity.badRequest()
-        // .body(Map.of("error", "Invalid email. Please enter a valid email address."));
+        // .body(Map.of("message", "Invalid email. Please enter a valid email
+        // address."));
         // }
         try {
             otpService.generateAndSendOtp(email, OTPPurpose.SIGN_UP);
-            return ResponseEntity.ok(Map.of("message", "OTP sent successfully to " + email));
+            return ResponseEntity.ok(Map.of("message", "OTP sent successfully to " + email + ". Valid for 5 minutes"));
         } catch (ResponseStatusException ex) {
             return ResponseEntity.status(ex.getStatusCode())
-                    .body(Map.of("error", ex.getReason()));
+                    .body(Map.of("message", ex.getReason()));
         } catch (MailException ex) {
             return ResponseEntity.status(HttpStatus.BAD_GATEWAY)
-                    .body(Map.of("error", "Failed to send OTP to the provided email. Please check the email address."));
+                    .body(Map.of("message",
+                            "Failed to send OTP to the provided email. Please check the email address."));
         } catch (Exception ex) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "An unexpected error occurred while sending OTP."));
+                    .body(Map.of("message", "An unexpected error occurred while sending OTP."));
         }
     }
 
@@ -165,15 +183,18 @@ public class AuthController {
         String otp = request.get("otp");
 
         if (email == null || otp == null)
-            return ResponseEntity.badRequest().body(Map.of("error", "Email and OTP are required"));
+            return ResponseEntity.badRequest().body(Map.of("message", "Email and OTP are required"));
 
         boolean verified = otpService.verifyOtp(email, otp, OTPPurpose.SIGN_UP);
-
-        if (verified)
-            return ResponseEntity.ok(Map.of("message", "OTP verified successfully"));
-        else
+        if (verified) {
+            String token = jwtUtil.generateOTPToken(email, OTPPurpose.SIGN_UP);
+            return ResponseEntity.ok(
+                    Map.of(
+                            "message", "OTP verified successfully",
+                            "otpToken", token));
+        } else
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body(Map.of("error", "Invalid Email or expired OTP"));
+                    .body(Map.of("message", "Invalid Email or expired OTP"));
     }
 
     @PostMapping("/forgot-password")
@@ -181,47 +202,64 @@ public class AuthController {
         String email = request.get("email");
         if (email == null || email.isBlank()) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "Email is required"));
+                    .body(Map.of("message", "Email is required"));
         }
         if (!userService.existsByEmail(email)) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "No user found"));
+                    .body(Map.of("message", "No user found"));
         }
         try {
             otpService.generateAndSendOtp(email, OTPPurpose.FORGOT_PASSWORD);
-            return ResponseEntity.ok(Map.of("message", "OTP sent successfully to " + email));
+            return ResponseEntity.ok(Map.of("message", "OTP sent successfully to " + email + ". Valid for 5 minutes."));
         } catch (ResponseStatusException ex) {
             return ResponseEntity.status(ex.getStatusCode())
-                    .body(Map.of("error", ex.getReason()));
+                    .body(Map.of("message", ex.getReason()));
         } catch (MailException ex) {
             return ResponseEntity.status(HttpStatus.BAD_GATEWAY)
-                    .body(Map.of("error", "Failed to send OTP to the provided email. Please check the email address."));
+                    .body(Map.of("message",
+                            "Failed to send OTP to the provided email. Please check the email address."));
         } catch (Exception ex) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "An unexpected error occurred while sending OTP."));
+                    .body(Map.of("message", "An unexpected error occurred while sending OTP."));
         }
     }
 
     @PostMapping("/verify-forgot-otp")
     public ResponseEntity<?> verifyForgotOtp(@RequestBody Map<String, String> request) {
-        String email = request.get("email");
-        String otp = request.get("otp");
+        try {
+            String email = request.get("email");
+            String otp = request.get("otp");
 
-        if (email == null || otp == null)
-            return ResponseEntity.badRequest().body(Map.of("error", "Email and OTP are required"));
+            if (email == null || email.isBlank() || otp == null || otp.isBlank()) {
+                return ResponseEntity.badRequest()
+                        .body(Map.of("message", "Email and OTP are required"));
+            }
 
-        boolean verified = otpService.verifyOtp(email, otp, OTPPurpose.FORGOT_PASSWORD);
+            boolean verified = otpService.verifyOtp(email, otp, OTPPurpose.FORGOT_PASSWORD);
 
-        if (verified)
-            return ResponseEntity.ok(Map.of("message", "OTP verified successfully"));
-        else
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body(Map.of("error", "Invalid Email or expired OTP"));
+            if (!verified) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                        .body(Map.of("message", "Invalid or expired OTP"));
+            }
+
+            String token = jwtUtil.generateOTPToken(email, OTPPurpose.FORGOT_PASSWORD);
+
+            return ResponseEntity.ok(
+                    Map.of(
+                            "message", "OTP verified successfully",
+                            "otpToken", token));
+        } catch (ResponseStatusException ex) {
+            return ResponseEntity.status(ex.getStatusCode())
+                    .body(Map.of("message", ex.getReason()));
+        } catch (Exception ex) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(Map.of("message", "An unexpected error occurred while verifying OTP"));
+        }
     }
 
     @PostMapping("/reset-password")
-    public ResponseEntity<?> resetPassword(@Valid @RequestBody ResetPasswordRequestDTO request, BindingResult result) {
-
+    public ResponseEntity<?> resetPassword(@RequestHeader(value = "x-otp-verify-token", required = false) String token,
+            @Valid @RequestBody ResetPasswordRequestDTO request, BindingResult result) {
         if (result.hasErrors()) {
 
             Map<String, String> errors = new HashMap<>();
@@ -229,30 +267,37 @@ public class AuthController {
 
             return ResponseEntity.badRequest().body(Map.of("errors", errors));
         }
+        if (token == null || token.isBlank()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body(Map.of("message", "Invalid Request. Please verify your email first"));
+        }
 
         String email = request.getEmail();
         String password = request.getPassword();
-        if (!otpService.isEmailVerified(email, OTPPurpose.FORGOT_PASSWORD)) {
 
-            return ResponseEntity.badRequest().body(
-                    Map.of("error", "Please verify your email before reset password."));
+        String emailFromToken = jwtUtil.validateOTPToken(token, OTPPurpose.FORGOT_PASSWORD);
+
+        if (emailFromToken == null || !emailFromToken.equals(email)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body(Map.of("message", "Invalid email"));
         }
-
+        if (!otpService.isEmailVerified(email, OTPPurpose.FORGOT_PASSWORD)) {
+            return ResponseEntity.badRequest().body(
+                    Map.of("message", "Please verify your email before reset password."));
+        }
         try {
-
             User savedUser = userService.setPassword(email, password);
-
             otpService.deleteByEmailAndPurpose(savedUser.getEmail(), OTPPurpose.FORGOT_PASSWORD);
             mailService.sendPasswordResetMail(savedUser.getEmail(), savedUser.getFirstName());
             return ResponseEntity.ok(Map.of("message", "Password Succesfully updated"));
 
         } catch (UsernameNotFoundException ex) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "User not found"));
+                    .body(Map.of("message", "User not found"));
 
         } catch (Exception ex) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "An unexpected error occurred", "message", ex.getMessage()));
+                    .body(Map.of("status", "An unexpected error occurred", "message", ex.getMessage()));
         }
     }
